+++ docker/.env
@@ -33,7 +33,7 @@ AM_CERTS_GRPC_CLIENT_TLS=
 AM_CERTS_GRPC_CA_CERTS=
 AM_CERTS_INSTANCE_ID=
 AM_CERTS_RELEASE_TAG=latest
-AM_CERTS_AGENT_TOKEN=
+AM_CERTS_TOKEN=
 
 ## OpenBao PKI
 AM_CERTS_OPENBAO_HOST=http://certs-openbao:8200
